moved some files in demo project
