+++ demo/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".LoginActivity">
+        <activity android:name=".auth.LoginActivity">
             <intent-filter>
                 <action android:name="eu.unicate.auth.action.AUTH"/>
                 <category android:name="android.intent.category.DEFAULT"/>
@@ -25,7 +25,7 @@
         </activity>
 
         <service
-            android:name=".DemoLoginService"
+            android:name=".auth.DemoLoginService"
             android:process=":auth"
             android:exported="false">
             <intent-filter>
