+++ demo/src/main/java/eu/unicate/retroauth/demo/GithubService.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import eu.unicate.retroauth.annotations.Authenticated;
 import eu.unicate.retroauth.annotations.Authentication;
-import eu.unicate.retroauth.demo.github.Email;
+import eu.unicate.retroauth.demo.auth.github.model.Email;
 import retrofit.Callback;
 import retrofit.http.GET;
 import rx.Observable;
