+++ demo/src/main/java/eu/unicate/retroauth/demo/auth/DemoLoginService.java
@@ -1,8 +1,9 @@
-package eu.unicate.retroauth.demo;
+package eu.unicate.retroauth.demo.auth;
 
 import android.content.Context;
 
 import eu.unicate.retroauth.AuthenticationService;
+import eu.unicate.retroauth.demo.R;
 
 public class DemoLoginService extends AuthenticationService {
 	@Override
