+++ demo/src/main/java/eu/unicate/retroauth/demo/auth/github/GithubAuthenticationService.java
@@ -1,10 +1,15 @@
-package eu.unicate.retroauth.demo.github;
+package eu.unicate.retroauth.demo.auth.github;
 
+import eu.unicate.retroauth.demo.auth.LoginActivity;
+import eu.unicate.retroauth.demo.auth.github.model.AccessToken;
 import retrofit.http.GET;
 import retrofit.http.Headers;
 import retrofit.http.Query;
 import rx.Observable;
 
+/**
+ * This api defines a method to get the AccessToken. It is used in the {@link LoginActivity}
+ */
 public interface GithubAuthenticationService {
 	@Headers({"Accept: application/json"})
 	@GET("/login/oauth/access_token")
