+++ demo/src/main/java/eu/unicate/retroauth/demo/auth/github/model/Email.java
@@ -1,4 +1,4 @@
-package eu.unicate.retroauth.demo.github;
+package eu.unicate.retroauth.demo.auth.github.model;
 
 public class Email {
 	public String email;
