+++ demo/src/main/java/eu/unicate/retroauth/demo/auth/github/model/GithubUser.java
@@ -1,0 +1,5 @@
+package eu.unicate.retroauth.demo.auth.github.model;
+
+public class GithubUser {
+	public String login;
+}
