--- demo/src/main/java/eu/unicate/retroauth/demo/github/GithubUser.java
@@ -1,5 +1,0 @@
-package eu.unicate.retroauth.demo.github;
-
-public class GithubUser {
-	public String login;
-}
